* origin/master:
  Bump Docker API version to match available
  Remove now-unnecessary TERRAFORM overrides
  Handle Packet worker server count greater than 1
  Ensure hostname template exists for Packet NAT
  Assert correct commands are run
  Remove upstart-related bits
  Ensure privnet bits run on boot
  Minor touchups upon reviewing summary diff in #416
  Rework container cleanup bits
  Container health script/cron touchups
  Use container health scripts on Packet, too

+++ modules/packet_worker/cloud-init.bash
@@ -15,13 +15,16 @@ main() {
   : "${ETCDIR:=/etc}"
   : "${RUNDIR:=/var/tmp/travis-run.d}"
   : "${VARTMP:=/var/tmp}"
+  : "${VARLIBDIR:=/var/lib}"
 
   export DEBIAN_FRONTEND=noninteractive
   chown nobody:nogroup "${VARTMP}"
   chmod 0777 "${VARTMP}"
 
   mkdir -p "${RUNDIR}"
-  echo "___INSTANCE_ID___-$(hostname)" >"${RUNDIR}/instance-hostname.tmpl"
+  if [[ ! -f "${RUNDIR}/instance-hostname.tmpl" ]]; then
+    echo "___INSTANCE_ID___-$(hostname)" >"${RUNDIR}/instance-hostname.tmpl"
+  fi
 
   for substep in \
     tfw \
@@ -83,6 +86,8 @@ __setup_networking() {
   done
 
   apt-get install -yqq iptables-persistent
+
+  "${VARLIBDIR}/cloud/scripts/per-boot/00-travis-packet-privnet-setup" || true
 }
 
 __setup_raid() {
